backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
python-multipart==0.0.6
